Throttle concurrency
Before this commit, work on a job would always start as soon as it was
accepted, regardless of whether the system was overloaded.  This
prevents too many threads from being active at one time -- new work will
still be accepted immediately but the execution will block until the
load is reasonable.