New and better errors format
Job.Errors is now a map from filenames to error strings.  This makes the
posted responses more parsable than encoding the filename in the error
string.

As a result of this change, initializing a new Job becomes somewhat more
complicated, since Errors is a map and must be allocated before use.
The server package shouldn't have to do this, so we have a new Job
constructor that takes JSON.

Since we have a constructor, make jobs initialize their own s3
connection instead of combiner, who shouldn't know that implemenation
detail.