Remove testify from main.go
Thought this would be a harmless way to ensure that this test
dependency gets fetched by `go get`, but as it turns out, doing
this also includes the test package as a dependency, which builds
a bunch of test options into the list of available flags and
unnecessarily bloats the binary.  Instead, just tell travis to get
testify before running the tests.