Disable throttling
This is a flawed approach.  Consider what happens when the number of
goroutines exceeds the threshold before this code is reached -- for
example if a large number of connections are made in a short duration.
The workers deadlock and never make any progress.  We need a smarter way
to do this.