Support embedded documents in requests
This adds a new optional member of the Document schema: `data`.
If present, it should be a base64-encoded zlib-compressed copy of the
file to be combined.  If this attribute is present, the 'Name'
attribute, normally used to locate the file on amazon, will be ignored
in favor of the directly-passed data.

The ruby code for generating the file data looks like:

  Base64.encode64(Zlib::deflate( file_content ))