Use Documents instead of strings representing them
This fixes up some debt related to the fact that when combiner was
originally written there was no object to represent a document.  Now
there is, so we should be passing them around instead of strings.  Also,
get rid of the Stat type and move its functionality (representing
document metadata) to Document itself.  In the mean time, eliminate some
timing and logging code for cleanliness, possibly to be added back later
if necessary.