Jobs remember their working directory
This information used to be stored in various places throughout
combiner, but it makes more sense for it to be a field on the Job.
Among other things, it allows Documents to know how to generate their
own paths.  This lets us change cpdf.Merge() so that it can operate
directly on a job.