Abort processing doc if metadata extraction fails
We should not ignore the failure to extract a pagecount or bookmark list
from a document, as it is our best indication that a PDF is malformed.
Instead just add it to the errors list and only attempt to combine
documents that successfully downloaded and validated by not blowing up
cpdf when it tried to extract metadata.