Loosen scale-down policy
There's no point in killing a fresh server if the load that spawned it
drops off for only 10 minutes.  The threshold should be higher, since
the cost of the server is incurred for a full hour anyway.  Start
dropping servers after 30 minutes of no load, and wait 20 minutes
between scaledown events.