Modify scale alarms to know about SQS
Since we overflow jobs to SQS once local queues reach a given threshold,
we can no longer use local queue size as a metric for when to scale up.
We can use the size of the SQS queue, but SQS reporting is only
refreshed in cloudwatch every 5 minutes.  The most responsive metric is
our custom excess one that tracks the net job count every minute.